+++ packages/components/descriptions/__test__/descriptions.test.tsx
@@ -1,0 +1,7 @@
+import { describe, expect, test } from 'vitest'
+
+describe('descriptions/index.vue', () => {
+  test('render test', async () => {
+    expect('').includes('')
+  })
+})