+++ packages/play/src/__nav.vue
@@ -1,6 +1,8 @@
 <template>
   <div>
     <router-link class="link" to="/date-picker">date-picker</router-link>
+    <router-link class="link" to="/descriptions">descriptions</router-link>
+    <router-link class="link" to="/field-item">field-item</router-link>
     <router-link class="link" to="/dialog">dialog</router-link>
     <router-link class="link" to="/image-preview">image-preview</router-link>
     <router-link class="link" to="/pagination">pagination</router-link>