+++ packages/plus-pro-components/component.ts
@@ -6,6 +6,7 @@ import { PlusImagePreview } from '@plus-pro-components/components/image-preview'
 import { PlusTable } from '@plus-pro-components/components/table'
 import { PlusRadio } from '@plus-pro-components/components/radio'
 import { PlusDatePicker } from '@plus-pro-components/components/date-picker'
+import { PlusDescriptions } from '@plus-pro-components/components/descriptions'
 
 const plugins: DefineComponent[] = [
   PlusDialog,
@@ -13,7 +14,8 @@ const plugins: DefineComponent[] = [
   PlusImagePreview,
   PlusTable,
   PlusRadio,
-  PlusDatePicker
+  PlusDatePicker,
+  PlusDescriptions
 ] as unknown as DefineComponent[]
 
 export default plugins